Fixes breakages after the API change for the CdiPane. Refactors duplicate code for creating the Cdi configuration window.
